requirements.txt
altgraph==0.10.2
appnope==0.1.0
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
bdist-mpkg==0.5.0
BeautifulSoup==3.2.1
beautifulsoup4==4.3.2
bonjour-py==0.3
boto==2.40.0
bottle==0.12.10
bson==0.4.3
bz2file==0.98
cached-property==1.2.0
certifi==2016.2.28
chardet==2.3.0
codegen==1.0
contextlib2==0.5.3
cybox==2.0.0.1
cycler==0.10.0
damn-simple-jsonrpc-server==0.4.2
damn-simple-jsonrpc-server-django==0.2.1
dask==0.9.0
db.py==0.4.4
decorator==4.0.9
design-patterns==0.2.0
Django==1.10.2
django-ipython==0.1.1
django-simple-jsonschema==0.0.5
dpkt==1.8.8
elasticsearch==1.6.0
Flask==0.10.1
Flask-Login==0.3.2
functools32==3.2.3.post2
future==0.15.2
futures==3.0.5
gensim==0.12.4
geolocation-python==0.2.2
gnureadline==6.3.3
googlemaps==2.4.5
gunicorn==19.6.0
html5lib==0.9999999
httpretty==0.8.10
ipython==4.2.0
ipython-db==1.0
ipython-genutils==0.1.0
ipython-pip==0.3
itsdangerous==0.24
Jinja2==2.8
jsonlib==1.6.1
jsonlib2==1.5.2
jsonschema==2.5.1
lxml==3.6.4
macholib==1.5.1
maec==4.0.0
Mako==1.0.4
MarkupSafe==0.23
matplotlib==1.5.1
modulegraph==0.10.4
mpld3==0.3
mpmath==0.19
myNaiveBayes==1.0.0
naive-bayes==0.1.1
NaiveBayes==1.0.0
naiveBayesClassifier==0.1.3
networkx==1.11
nltk==3.0.2
nltk-tgrep==1.0.6
numpy==1.11.0
nxpydot==0.1
oauthlib==1.0.3
Orange==2.7.8
Orange-Canvas-Core==0.0.7
Orange-Multitarget==0.9.3
Orange-Network==0.3.5
Orange-Reliability==0.2.14
Orange-Text==1.2a1
Orange-Textable==1.5.2
Orange3-Text==0.1.10
pandas==0.15.2
pathlib2==2.1.0
Pattern==2.6
pattern-gen==0.1.0
Pattern-Generator==1.0.1
patterns==0.3
pefile==2016.3.28
pexpect==4.0.1
pickleshare==0.7.2
Pillow==3.2.0
prettytable==0.7.2
ptyprocess==0.5.1
py-pattern-matching==0.1.1
py2app==0.7.3
pybars3==0.9.1
pydeep==0.2
pydot==1.0.28
pydot-ng==1.0.0
pydot3==1.0.8
pydot3k==1.0.17
pydotplus==2.0.2
pylab==0.1.3
PyMeta3==0.5.1
pymongo==3.3.0
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.13.1
pyparsing==2.1.1
pypi-json==0.0.0
python-crfsuite==0.8.1
python-dateutil==2.5.3
python-magic==0.4.12
python-patterns==0.0.1
python-twitter==2.2
pytz==2016.7
pyzmq==15.2.0
regex==2015.3.18
requests==2.8.1
requests-oauthlib==0.6.1
scikit-image==0.12.3
scikit-learn==0.17.1
scipy==0.17.0
seaborn==0.7.0
setuptools-git==1.1
simple-jsonrpc==0.3.4
simplegeneric==0.8.1
simplejson==3.8.2
singledispatch==3.4.0.3
six==1.10.0
smart-open==1.3.2
SQLAlchemy==1.1.1
sympy==1.0
toolz==0.7.4
tornado==4.3
traitlets==4.2.1
tweepy==3.5.0
urllib3==1.15.1
WebPattern==0.0.3
Werkzeug==0.11.9
xattr==0.6.4
yara==1.7.7
zope.interface==4.1.1
